Create separate files for dtos JC-644
