+++ libs/core/dtos/aired.dto.ts
@@ -1,0 +1,4 @@
+export type AiredDto = Readonly<{
+	start: string | null;
+	end: string | null;
+}>;
