+++ libs/core/dtos/anime-status.dto.ts
@@ -1,0 +1,5 @@
+export enum AnimeStatusDto {
+	Airing = 'AIRING',
+	Finished = 'FINISHED',
+	NotYetAired = 'NOT_YET_AIRED',
+}
