+++ libs/core/dtos/anime-type.dto.ts
@@ -1,0 +1,10 @@
+export enum AnimeTypeDto {
+	Movie = 'MOVIE',
+	Music = 'MUSIC',
+	ONA = 'ONA',
+	OVA = 'OVA',
+	Promotional = 'PROMOTIONAL_VIDEOS',
+	Special = 'SPECIAL',
+	TV = 'TV',
+	Unknown = 'UNKNOWN',
+}
