+++ libs/core/dtos/anime.dto.ts
@@ -1,3 +1,7 @@
+import { AiredDto } from './aired.dto';
+import { AnimeStatusDto } from './anime-status.dto';
+import { AnimeTypeDto } from './anime-type.dto';
+
 /** Anime DTO. */
 export type AnimeDto = Readonly<{
 	id: number;
@@ -14,25 +18,3 @@ export type AnimeDto = Readonly<{
 	studios: readonly number[];
 	genres: readonly number[];
 }>;
-
-export type AiredDto = Readonly<{
-	start: string | null;
-	end: string | null;
-}>;
-
-export enum AnimeStatusDto {
-	Airing = 'AIRING',
-	Finished = 'FINISHED',
-	NotYetAired = 'NOT_YET_AIRED',
-}
-
-export enum AnimeTypeDto {
-	Movie = 'MOVIE',
-	Music = 'MUSIC',
-	ONA = 'ONA',
-	OVA = 'OVA',
-	Promotional = 'PROMOTIONAL_VIDEOS',
-	Special = 'SPECIAL',
-	TV = 'TV',
-	Unknown = 'UNKNOWN',
-}
